requirements.txt
discord.py==1.7.3
openai==0.27.0
beautifulsoup4==4.10.0
requests==2.26.0
